Move schema generator functions to separate schema package
