+++ integtest/consumer_test.go
@@ -16,6 +16,7 @@ import (
 	. "go.dataddo.com/pgq"
 	"go.dataddo.com/pgq/internal/pg"
 	"go.dataddo.com/pgq/internal/require"
+	"go.dataddo.com/pgq/schema"
 )
 
 func TestConsumer_Run_graceful_shutdown(t *testing.T) {
@@ -26,11 +27,11 @@ func TestConsumer_Run_graceful_shutdown(t *testing.T) {
 
 	db := openDB(t)
 	queueName := t.Name()
-	_, _ = db.ExecContext(ctx, GenerateDropTableQuery(queueName))
-	_, err := db.ExecContext(ctx, GenerateCreateTableQuery(queueName))
+	_, _ = db.ExecContext(ctx, schema.GenerateDropTableQuery(queueName))
+	_, err := db.ExecContext(ctx, schema.GenerateCreateTableQuery(queueName))
 	t.Cleanup(func() {
 		db := openDB(t)
-		_, err := db.ExecContext(ctx, GenerateDropTableQuery(queueName))
+		_, err := db.ExecContext(ctx, schema.GenerateDropTableQuery(queueName))
 		require.NoError(t, err)
 		err = db.Close()
 		require.NoError(t, err)
