+++ integtest/publisher_test.go
@@ -11,6 +11,7 @@ import (
 	"go.dataddo.com/pgq"
 	pgutils "go.dataddo.com/pgq/internal/pg"
 	"go.dataddo.com/pgq/internal/require"
+	"go.dataddo.com/pgq/schema"
 )
 
 func TestPublisher(t *testing.T) {
@@ -58,11 +59,11 @@ func TestPublisher(t *testing.T) {
 				require.NoError(t, err)
 			})
 			queueName := t.Name()
-			_, _ = db.ExecContext(ctx, pgq.GenerateDropTableQuery(queueName))
-			_, err := db.ExecContext(ctx, pgq.GenerateCreateTableQuery(queueName))
+			_, _ = db.ExecContext(ctx, schema.GenerateDropTableQuery(queueName))
+			_, err := db.ExecContext(ctx, schema.GenerateCreateTableQuery(queueName))
 			require.NoError(t, err)
 			t.Cleanup(func() {
-				_, err := db.ExecContext(ctx, pgq.GenerateDropTableQuery(queueName))
+				_, err := db.ExecContext(ctx, schema.GenerateDropTableQuery(queueName))
 				require.NoError(t, err)
 			})
 			d := pgq.NewPublisher(db, tt.publisherOpts...)
